Initial commit
This is BookCover-DL, a simple GUI application for downloading book covers
based on provided URL of supported service or ISBN of the book.

Copyright (C) 2024 Patryk Miś

This program is free software: you can redistribute it and/or modify
it under the terms of the GNU General Public License as published by
the Free Software Foundation, either version 3 of the License, or
(at your option) any later version.

This program is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
GNU General Public License for more details.

You should have received a copy of the GNU General Public License
along with this program.  If not, see <https://www.gnu.org/licenses/>.

Signed-off-by: Patryk Miś <foss@patrykmis.com>